+++ R/helpers_easystats.R
@@ -13,7 +13,7 @@ tidy_model_parameters <- function(model, ...) {
     select(-matches("Difference")) %>%
     standardize_names(style = "broom") %>%
     rename_all(~ gsub("cramers.", "", .x)) %>%
-    rename_all(.funs = recode, "bayes.factor" = "bf10") %>%
+    rename_with(recode, bayes.factor = "bf10") %>%
     tidyr::fill(matches("^prior|^bf"), .direction = "updown") %>%
     mutate(across(matches("bf10"), ~ log(.x), .names = "log_e_{.col}"))
 
@@ -26,7 +26,7 @@ tidy_model_parameters <- function(model, ...) {
     df_r2 <- performance::r2_bayes(model, average = TRUE, verbose = FALSE, ci = stats_df$conf.level[[1]]) %>%
       as_tibble() %>%
       standardize_names(style = "broom") %>%
-      rename("estimate" = "r.squared") %>%
+      rename(estimate = r.squared) %>%
       filter(if_any(matches("component"), ~ (.x == "conditional")))
 
     # remove estimates and CIs and use R2 data frame instead