+++ NEWS.md
@@ -1,3 +1,7 @@
+# statsExpressions 1.5.2
+
+- Fixes failures on R-devel due to changes in the formula interface for a paired test.
+
 # statsExpressions 1.5.1
 
 ## MAJOR CHANGES